Raport fix

+++ BudgetOrganizer/Models/OperationModel/OperationByCategoryReportDTO.cs
@@ -5,6 +5,6 @@ namespace BudgetOrganizer.Models.OperationModel
     public class OperationByCategoryReportDTO
     {
         public GetCategoryDTO Category { get; set; }
-        public int Sum { get; set; } = 0;
+        public decimal Sum { get; set; } = 0;
     }
 }
